==> Dolgos_PDOs_Early/4_clustering.Rmd <==
---
title: '4.Dolgos early passage PDOs : clustering'
author: "Romuald Parmentier"
date: "2024-10-09"
output: html_document
---

```{r Load libraries and create output folder, message = F}

# Libraries
library(Seurat)
library(SingleCellExperiment)
library(scater)
library(ggplot2)
library(dplyr)
library(bluster)
library(scran)

# Functions and palettes
source(paste0(github_dir, "/Dolgos_Custom_Functions.R"))
source(paste0(github_dir, "/Dolgos_Custom_Color_Palettes.R"))

# Create output path
out_path = create_exp_folder(
  github_dir = github_dir,
  samples_ID = "Dolgos_PDOs_Early",
  exp = "4_clustering"
)

```

# Prepare the data

```{r Load the files}

# Load rescaled sce after cell cycle annotation 
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Dolgos_PDOs_Early",
  prev_exp = "3_cycle_annotation", 
  pattern = "sce_comb_cycle_annotated.rds") # Sce after batch correction for visuaoization

sce_comb = readRDS(file = file_path)

```

# Perform clustering

```{r walktrap clustering with different K}

k_ids = c(10, 20, 30)

# For loop to calculate clustering information assignation
for(k_id in k_ids) {
  
  print(paste("Clustering with k =", k_id, "ongoing"))
  
  clustering = clusterCells(
    sce_comb,
    use.dimred = "MNN.1-20",
    BLUSPARAM = bluster::SNNGraphParam(
      k = k_id,
      type = "rank",
      cluster.fun = "walktrap")
  )
  
  # Create a column in colData with the type of clustering and the associated parameter
  sce_comb[[paste0("walktrap_", k_id)]] = clustering
  
}

```

```{r Malignancy assignation acording cluster}

sce_comb$Malignancy = ifelse(
  test = sce_comb$walktrap_30 %in% c(1,8),
  yes = "Tumor",
  no = ifelse(
    test = sce_comb$walktrap_30 == 4,
    yes = "Transitioning",
    no = ifelse(
      test = sce_comb$walktrap_30 == 3,
      yes = "Club",
      no = ifelse(
        test = sce_comb$walktrap_30 %in% c(2,5,6,7),
        yes = "Basal/Hillock",
          no = "Other")
    )
  )
)

```

# Visualize clustering

```{r Make a tibble of metadata}

# get metadata
md <- as_tibble(colData(sce_comb))

# get embeddings from 2 first PC
coords <- as_tibble(reducedDim(sce_comb, "UMAP_on_MNN.1-20"))[,1:2]
colnames(coords) = c("UMAP_1","UMAP_2")

# combine dataframes
md <- cbind(md, coords)

```

```{r Facette UMAP colord with cluster number according walktrap k values}

k_ids = c(10, 20, 30)  # Define clustering IDs
plot_id = 1            # Initialize plot ID counter
plot_list = list()     # Initialize empty list to store plots

for(k_id in k_ids) {
  
  clustering = paste0("walktrap_", k_id)  # Construct the clustering column name
  
  # Create ggplot with rasterized points for each clustering
  plot <- ggplot() +
    # Rasterize the geom_point layer
    ggrastr::rasterise(
      geom_point(
        data = md,  # DataFrame containing UMAP coordinates and clustering info
        aes(x = UMAP_1, y = UMAP_2, fill = .data[[clustering]]),  # Use the dynamic clustering column
        alpha = 0.7, size = 1.3, 
        shape = 21, color = "gray10", stroke = 0.2  # Lighter stroke color
      ), 
      dpi = 300, scale = 1  # Higher DPI and scale for better resolution
    ) +
    scale_fill_manual(values = pal_walktrap_cluster) +  # Use predefined color palette
    theme_classic() +  # Clean background
    guides(
      fill = "none",  # Disable legend for fill
      alpha = "none", 
      size = "none"
    ) +
    ggtitle(label = "Dolgos 2024 Early Passage PDO: Walktrap clustering",
            subtitle = paste0("Clustering param: NN = ", k_id)  # Dynamic subtitle
    )
  
  plot_list[[plot_id]] = plot  # Add the plot to the list
  plot_id = plot_id + 1        # Increment plot ID
}

# Arrange plots in a multi-page layout (2 plots per page)
multiple_page_layout <- gridExtra::marrangeGrob(
  grobs = plot_list,
  nrow = 2, ncol = 1  # Define layout (2 rows, 1 column per page)
)

# Save the arranged plots to a multi-page PDF
ggsave(
  plot = multiple_page_layout,
  filename = paste0(out_path,time_stamp(),"plots_UMAP_Facetted_walktrap_clustering_10-30.pdf"),
  device = "pdf",
  width = 210, height = 297, units = "mm"
)

```

```{r Single UMAP colored with clustering walktrap 30 = Figure 3.B}

# Calculate centroids for each cluster
centroids <- md %>%
  group_by(walktrap_30) %>%
  summarise(
    UMAP_1 = mean(UMAP_1),
    UMAP_2 = mean(UMAP_2),
    .groups = 'drop'
  )

plot <- ggplot() +
  # Rasterise the first layer
  ggrastr::rasterise(
    geom_point(
      data = md,
      aes(x = UMAP_1, y = UMAP_2, fill = walktrap_30), 
      alpha = 0.7, size = 1.3, 
      shape = 21,
      color = "gray10",
      stroke = 0.2),  # Lighter stroke color (adjust as needed)),# Ensure shape supports fill and color
    dpi = 300, scale = 1) +
  scale_fill_manual(values = pal_walktrap_cluster) +
  theme_classic() +
  guides(
    fill = "none",
    alpha = "none",
    size = "none"
  ) +
  # Add cluster numbers in the center of each cluster
  geom_text(
    data = centroids,
    aes(x = UMAP_1, y = UMAP_2, label = walktrap_30),
    color = "black",  # Color for the cluster numbers
    size = 7,  # Adjust size as needed
    fontface = "bold"
  )

ggsave(
  plot = plot,
  filename = paste0(out_path,time_stamp(),"plot_UMAP_Single_Walktrap_30.pdf"),
  device = "pdf",
  width = 10,
  height = 8,
)

```

```{r Single UMAP colored with Malignancy}

plot <- ggplot() +
  # Rasterise the first layer
  ggrastr::rasterise(
    geom_point(
      data = md,
      aes(x = UMAP_1, y = UMAP_2, fill = Malignancy), 
      alpha = 0.7, size = 1.3, 
      shape = 21,
      color = "gray10",
      stroke = 0.2),  # Lighter stroke color (adjust as needed)),# Ensure shape supports fill and color
    dpi = 300, scale = 1) +
  scale_fill_manual(values = pal_malignancy_type) +
  theme_classic() +
  guides(
    fill = "none",
    alpha = "none",
    size = "none"
  ) 

ggsave(
  plot = plot,
  filename = paste0(out_path,time_stamp(),"plot_UMAP_Single_Malignancy.pdf"),
  device = "pdf",
  width = 10,
  height = 8,
)

```

```{r Histogram Sample per cluster = Figure 3.C}

md_histo <- md %>%
  group_by(walktrap_30, Sample_Description) %>%
  summarise(count = n()) %>%
  ungroup()

md_histo = md_histo %>%
  group_by(walktrap_30) %>%
  mutate(total_cluster = sum(count)) %>%
  mutate(pct_sample_cluster =(count/total_cluster)*100)

# Create the bar plot
plot = ggplot(md_histo, aes(x = walktrap_30, y = pct_sample_cluster, fill = Sample_Description)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_manual(values = pal_Sample_Description) + 
  labs(title = "Percentages of sample per cluster", x = "Cluster", y = "Count") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) 

ggsave(plot,
       filename = paste0(out_path,time_stamp(),"plot_Histogram_Walktrap_30_Sample.pdf"),
       device = "pdf",
       width = 6,
       height = 6
)

```

```{r Histogram Culture Condition per cluster = Figure 3.D}

md_histo <- md %>%
  group_by(walktrap_30, Culture_Condition) %>%
  summarise(count = n()) %>%
  ungroup() %>% 
  tidyr::complete(walktrap_30, Culture_Condition, fill = list(count = 0))

md_histo = md_histo %>%
  group_by(walktrap_30) %>%
  mutate(total_cluster = sum(count)) %>%
  mutate(pct_sample_cluster = (count/total_cluster)*100)   

# Order walktrap cluster with ascending % of Matrigel cultures org nb
pct_Matrigel = md_histo %>%
  ungroup %>%
  dplyr::filter(Culture_Condition == "Matrigel") %>%
  dplyr::select(walktrap_30, pct_sample_cluster ) %>%
  arrange(pct_sample_cluster)

# Rearrange the cluster in the table
md_histo$walktrap_30 = factor(x = md_histo$walktrap_30, levels = pct_Matrigel$walktrap_30)

# Create the bar plot
plot = ggplot(md_histo, aes(x = walktrap_30, y = pct_sample_cluster, fill = Culture_Condition)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_manual(values = pal_Culture_Condition) + 
  labs(title = "Percentages of culture condition per cluster", x = "Cluster", y = "Count") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) 

ggsave(plot,
       filename = paste0(out_path,time_stamp(),"plot_Histogram_Walktrap_30_Culture_Condition.pdf"),
       device = "pdf",
       width = 6,
       height = 6
)

```

```{r Histogram cell cycle per cluster = Sup.Fig3.A}

md_histo <- md %>%
  group_by(walktrap_30, cell_cycle_phase) %>%
  summarise(count = n()) %>%
  ungroup()

md_histo = md_histo %>%
  group_by(walktrap_30) %>%
  mutate(total_cluster = sum(count)) %>%
  mutate(pct_sample_cluster =(count/total_cluster)*100)

# Create the bar plot
plot = ggplot(md_histo, aes(x = walktrap_30, y = pct_sample_cluster, fill = cell_cycle_phase)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_manual(values = pal_cell_cycle_phase) + 
  labs(title = "Percentages of cell per cluster", x = "Cluster", y = "Count") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) 

ggsave(plot,
       filename = paste0(out_path,time_stamp(),"plot_Histogram_Walktrap_30_Cell_Cycle.pdf"),
       device = "pdf",
       width = 6,
       height = 6
)

```

```{r Facetted Violin plot of QC stats per cluster}

QC_stat = c(
  "Percent Mitochondrial reads per cell" = "subsets_mitochondrial_reads_percent",
  "Number of gene per cell" = "detected",
  "Number of UMI per cell" = "sum",
  "Doublet finder score" = "scDblFinder.score"
)

plot_id = 1
plot_list = list()

for(qc in QC_stat){
  
  plot <- ggplot(md, aes(x = walktrap_30, y = .data[[qc]], fill = walktrap_30)) +
    geom_violin(alpha = 0.8, position = position_dodge(width = 0.8), scale = "width") +
    geom_boxplot(width = 0.1, position = position_dodge(width = 0.8), alpha = 0.5) +
    scale_fill_manual(values = pal_walktrap_cluster, name = "walktrap_30") +
    guides(fill = "none") +
    theme_minimal() + 
    theme(
      axis.text.x = element_text(angle = 45, hjust = 1),
      axis.title = element_text(size = 12),  # Customize axis title appearance
      axis.text = element_text(size = 10),    # Customize axis text appearance
      axis.line = element_line(size = 0.5),   # Customize axis line appearance
      axis.ticks = element_line(size = 0.5)) +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    labs(
      title = element_text(names(QC_stat)[plot_id]),
      x = element_blank())
  
  plot_list[[plot_id]] = plot  # Add the plot to the list
  plot_id = plot_id + 1        # Increment plot ID
}

# Arrange plots in a multi-page layout (2 plots per page)
multiple_page_layout <- gridExtra::marrangeGrob(
  grobs = plot_list,
  nrow = 2, ncol = 1  # Define layout (2 rows, 1 column per page)
)

# Save the arranged plots to a multi-page PDF
ggsave(
  plot = multiple_page_layout,
  filename = paste0(out_path,time_stamp(),"plots_violin_Walktrap_30_QC_Stats.pdf"),
  device = "pdf",
  width = 210, height = 297, units = "mm"
)

```

# Export files

```{r Export clustered sce object}

# Clustering takes a lot of time, better to export sce object here

saveRDS(
  object = sce_comb,
  file = paste0(out_path,time_stamp(),"sce_comb_clustered.rds")
)

```
